+++ influxql/token.go
@@ -231,11 +231,9 @@ var keywords map[string]Token
 func init() {
 	keywords = make(map[string]Token)
 	for tok := keyword_beg + 1; tok < keyword_end; tok++ {
-		keywords[strings.ToUpper(tokens[tok])] = tok
 		keywords[strings.ToLower(tokens[tok])] = tok
 	}
 	for _, tok := range []Token{AND, OR} {
-		keywords[strings.ToUpper(tokens[tok])] = tok
 		keywords[strings.ToLower(tokens[tok])] = tok
 	}
 	keywords["true"] = TRUE
@@ -280,7 +278,7 @@ func tokstr(tok Token, lit string) string {
 
 // Lookup returns the token associated with a given string.
 func Lookup(ident string) Token {
-	if tok, ok := keywords[ident]; ok {
+	if tok, ok := keywords[strings.ToLower(ident)]; ok {
 		return tok
 	}
 	return IDENT